requirements.txt
hikari[speedups]==2.0.0.dev126
hikari-lightbulb==2.3.5
PyYAML==6.0